Git-related and administrative changes...
configure.ac
-------------

save autoconf variables for MAKE (== "make") and "MAKE_VERSION" (4.3)
these are used in mainMakefile to get the po-files

there were also some stray typos in this file "+" at the end of a line
and AC_FUNC_VPRINTF

maintMakefile
-------------

Start merging
http://git.savannah.gnu.org/cgit/make.git/tree/maintMakefile

The way po_files are retrieved has changed.

There are some stray files in lib/ that cause a problem in
po-generation, so that's been hacked around. (see FIXME) in the comment
of maintMakefile

.gitignores
-----------

Exclude dervative files, i.e those like src/config.h that get created in
the build procoess. A lot more file should be added, but this is a start.